Time: 0 ms (100%), Space: 22.7 MB (19.71%) - LeetHub

+++ 0026-remove-duplicates-from-sorted-array/0026-remove-duplicates-from-sorted-array.cpp
@@ -26,7 +26,7 @@ class Solution {
 public:
     int removeDuplicates(vector<int>& nums) {
         short sizeNums = nums.size();
-        short *indcies=new short[sizeNums];
+        int *indcies=new int[sizeNums];
         fill_n(indcies, sizeNums, -1);
         short start = 0;
         for (short i = sizeNums - 1; i > 0; i--)
